requirements.txt
streamlit>=1.32.0
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langchain-core>=0.2.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
